+++ app/handlers_flightrecord.go
@@ -9,6 +9,7 @@ import (
 	"github.com/go-chi/chi/v5"
 	"github.com/google/uuid"
 	"github.com/vsimakhin/web-logbook/internal/models"
+	"github.com/vsimakhin/web-logbook/internal/utils"
 )
 
 // HandlerFlightRecordByID shows flight record by UUID
@@ -26,7 +27,6 @@ func (app *application) HandlerApiFlightRecordByID(w http.ResponseWriter, r *htt
 
 // HandlerNightTime is a handler for calculating night time
 func (app *application) HandlerNightTime(w http.ResponseWriter, r *http.Request) {
-
 	var fr models.FlightRecord
 	var response models.JSONResponse
 
@@ -69,7 +69,6 @@ func (app *application) HandlerApiFlightRecordDelete(w http.ResponseWriter, r *h
 
 // HandlerApiFlightRecordNew creates a new flight record
 func (app *application) HandlerApiFlightRecordNew(w http.ResponseWriter, r *http.Request) {
-
 	var flightRecord models.FlightRecord
 	err := json.NewDecoder(r.Body).Decode(&flightRecord)
 	if err != nil {
@@ -100,12 +99,17 @@ func (app *application) HandlerApiFlightRecordNew(w http.ResponseWriter, r *http
 
 // HandlerApiFlightRecordUpdate updates the flight record
 func (app *application) HandlerApiFlightRecordUpdate(w http.ResponseWriter, r *http.Request) {
-
 	var flightRecord models.FlightRecord
 	err := json.NewDecoder(r.Body).Decode(&flightRecord)
 	if err != nil {
 		app.handleError(w, err)
 		return
+	}
+
+	// recalculate distance in case departure or arrival place has been changed
+	// but not if track log is present
+	if flightRecord.Track == nil {
+		flightRecord.Distance = app.db.Distance(flightRecord.Departure.Place, flightRecord.Arrival.Place)
 	}
 
 	err = app.db.UpdateFlightRecord(flightRecord)
@@ -115,4 +119,57 @@ func (app *application) HandlerApiFlightRecordUpdate(w http.ResponseWriter, r *h
 	}
 
 	app.writeOkResponse(w, "Flight Record has been updated")
+}
+
+// HandlerApiTrackLogNew is a handler for uploading track log
+func (app *application) HandlerApiTrackLogNew(w http.ResponseWriter, r *http.Request) {
+	uuid := chi.URLParam(r, "uuid")
+
+	var track [][]float64
+	err := json.NewDecoder(r.Body).Decode(&track)
+	if err != nil {
+		app.handleError(w, err)
+		return
+	}
+
+	distance := 0.0
+	for i := range len(track) - 1 {
+		distance += utils.Distance(track[i][0], track[i][1], track[i+1][0], track[i+1][1])
+	}
+
+	bTrack, err := json.Marshal(track)
+	if err != nil {
+		app.handleError(w, err)
+		return
+	}
+
+	err = app.db.UpdateFlightRecordTrack(uuid, distance, bTrack)
+	if err != nil {
+		app.handleError(w, err)
+		return
+	}
+
+	app.writeOkResponse(w, "Track Log uploaded")
+}
+
+// HandlerApiTrackLogReset is a handler for deleting track log
+func (app *application) HandlerApiTrackLogReset(w http.ResponseWriter, r *http.Request) {
+	uuid := chi.URLParam(r, "uuid")
+
+	fr, err := app.db.GetFlightRecordByID(uuid)
+	if err != nil {
+		app.handleError(w, err)
+		return
+	}
+
+	// recalculate distance
+	fr.Distance = app.db.Distance(fr.Departure.Place, fr.Arrival.Place)
+
+	err = app.db.UpdateFlightRecordTrack(uuid, fr.Distance, nil)
+	if err != nil {
+		app.handleError(w, err)
+		return
+	}
+
+	app.writeOkResponse(w, "Track Log reset")
 }