+++ internal/models/flightrecords.go
@@ -104,23 +104,28 @@ func (m *DBModel) GetFlightRecordByID(uuid string) (fr FlightRecord, err error) 
 	ctx, cancel := m.ContextWithDefaultTimeout()
 	defer cancel()
 
-	query := "SELECT uuid, date, m_date, departure_place, departure_time, " +
-		"arrival_place, arrival_time, aircraft_model, reg_name, " +
-		"se_time, me_time, mcc_time, total_time, day_landings, night_landings, " +
-		"night_time, ifr_time, pic_time, co_pilot_time, dual_time, instructor_time, " +
-		"sim_type, sim_time, pic_name, remarks " +
-		"FROM logbook_view WHERE uuid = ?"
+	query := `
+		SELECT 
+			uuid, date, m_date, departure_place, departure_time,
+			arrival_place, arrival_time, aircraft_model, reg_name,
+			se_time, me_time, mcc_time, total_time, day_landings, night_landings,
+			night_time, ifr_time, pic_time, co_pilot_time, dual_time, instructor_time,
+			sim_type, sim_time, pic_name, remarks, distance, track
+		FROM logbook_view 
+		WHERE uuid = ?`
 	row := m.DB.QueryRowContext(ctx, query, uuid)
 
-	if err = row.Scan(&fr.UUID, &fr.Date, &fr.MDate, &fr.Departure.Place, &fr.Departure.Time,
+	err = row.Scan(
+		&fr.UUID, &fr.Date, &fr.MDate, &fr.Departure.Place, &fr.Departure.Time,
 		&fr.Arrival.Place, &fr.Arrival.Time, &fr.Aircraft.Model, &fr.Aircraft.Reg,
 		&fr.Time.SE, &fr.Time.ME, &fr.Time.MCC, &fr.Time.Total, &fr.Landings.Day, &fr.Landings.Night,
 		&fr.Time.Night, &fr.Time.IFR, &fr.Time.PIC, &fr.Time.CoPilot, &fr.Time.Dual, &fr.Time.Instructor,
-		&fr.SIM.Type, &fr.SIM.Time, &fr.PIC, &fr.Remarks); err != nil {
+		&fr.SIM.Type, &fr.SIM.Time, &fr.PIC, &fr.Remarks, &fr.Distance, &fr.Track,
+	)
+	if err != nil {
 		return fr, err
 	}
-
-	// calculate distance
+	// process flight record
 	m.processFlightrecord(&fr)
 
 	// get previous and next records uuid
@@ -167,19 +172,21 @@ func (m *DBModel) UpdateFlightRecord(fr FlightRecord) error {
 	ctx, cancel := m.ContextWithDefaultTimeout()
 	defer cancel()
 
-	query := "UPDATE logbook SET " +
-		"date = ?, departure_place = ?, departure_time = ?, " +
-		"arrival_place = ?, arrival_time = ?, aircraft_model = ?, reg_name = ?, " +
-		"se_time = ?, me_time = ?, mcc_time = ?, total_time = ?, day_landings = ?, night_landings = ?, " +
-		"night_time = ?, ifr_time = ?, pic_time = ?, co_pilot_time = ?, dual_time = ?, instructor_time = ?, " +
-		"sim_type = ?, sim_time = ?, pic_name = ?, remarks = ? " +
-		"WHERE uuid = ?"
+	query := `
+		UPDATE logbook SET 
+			date = ?, departure_place = ?, departure_time = ?,
+			arrival_place = ?, arrival_time = ?, aircraft_model = ?, reg_name = ?,
+			se_time = ?, me_time = ?, mcc_time = ?, total_time = ?, day_landings = ?, night_landings = ?,
+			night_time = ?, ifr_time = ?, pic_time = ?, co_pilot_time = ?, dual_time = ?, instructor_time = ?,
+			sim_type = ?, sim_time = ?, pic_name = ?, remarks = ?, distance = ?
+		WHERE uuid = ?`
 	_, err := m.DB.ExecContext(ctx, query,
 		fr.Date, fr.Departure.Place, fr.Departure.Time,
 		fr.Arrival.Place, fr.Arrival.Time, fr.Aircraft.Model, fr.Aircraft.Reg,
 		fr.Time.SE, fr.Time.ME, fr.Time.MCC, fr.Time.Total, fr.Landings.Day, fr.Landings.Night,
 		fr.Time.Night, fr.Time.IFR, fr.Time.PIC, fr.Time.CoPilot, fr.Time.Dual, fr.Time.Instructor,
-		fr.SIM.Type, fr.SIM.Time, fr.PIC, fr.Remarks, fr.UUID,
+		fr.SIM.Type, fr.SIM.Time, fr.PIC, fr.Remarks, fr.Distance,
+		fr.UUID,
 	)
 	return err
 }
@@ -189,23 +196,26 @@ func (m *DBModel) InsertFlightRecord(fr FlightRecord) error {
 	ctx, cancel := m.ContextWithDefaultTimeout()
 	defer cancel()
 
-	query := "INSERT INTO logbook " +
-		"(uuid, date, departure_place, departure_time, " +
-		"arrival_place, arrival_time, aircraft_model, reg_name, " +
-		"se_time, me_time, mcc_time, total_time, day_landings, night_landings, " +
-		"night_time, ifr_time, pic_time, co_pilot_time, dual_time, instructor_time, " +
-		"sim_type, sim_time, pic_name, remarks) " +
-		"VALUES (?, ?, ?, ?, " +
-		"?, ?, ?, ?, " +
-		"?, ?, ?, ?, ?, ?, " +
-		"?, ?, ?, ?, ?, ?, " +
-		"?, ?, ?, ?)"
+	fr.Distance = m.Distance(fr.Departure.Place, fr.Arrival.Place)
+
+	query := `
+		INSERT INTO logbook 
+			(uuid, date, departure_place, departure_time,
+			arrival_place, arrival_time, aircraft_model, reg_name,
+			se_time, me_time, mcc_time, total_time, day_landings, night_landings,
+			night_time, ifr_time, pic_time, co_pilot_time, dual_time, instructor_time,
+			sim_type, sim_time, pic_name, remarks, distance)
+		VALUES (?, ?, ?, ?,
+		?, ?, ?, ?,
+		?, ?, ?, ?, ?, ?,
+		?, ?, ?, ?, ?, ?,
+		?, ?, ?, ?, ?)`
 	_, err := m.DB.ExecContext(ctx, query,
 		fr.UUID, fr.Date, fr.Departure.Place, fr.Departure.Time,
 		fr.Arrival.Place, fr.Arrival.Time, fr.Aircraft.Model, fr.Aircraft.Reg,
 		fr.Time.SE, fr.Time.ME, fr.Time.MCC, fr.Time.Total, fr.Landings.Day, fr.Landings.Night,
 		fr.Time.Night, fr.Time.IFR, fr.Time.PIC, fr.Time.CoPilot, fr.Time.Dual, fr.Time.Instructor,
-		fr.SIM.Type, fr.SIM.Time, fr.PIC, fr.Remarks,
+		fr.SIM.Type, fr.SIM.Time, fr.PIC, fr.Remarks, fr.Distance,
 	)
 	return err
 }
@@ -230,7 +240,7 @@ func (m *DBModel) GetFlightRecords() (flightRecords []FlightRecord, err error) {
 			arrival_place, arrival_time, aircraft_model, reg_name,
 			se_time, me_time, mcc_time, total_time, day_landings, night_landings,
 			night_time, ifr_time, pic_time, co_pilot_time, dual_time, instructor_time,
-			sim_type, sim_time, pic_name, remarks
+			sim_type, sim_time, pic_name, remarks, distance
 		FROM logbook_view
 		ORDER BY m_date desc, departure_time desc`)
 
@@ -241,11 +251,63 @@ func (m *DBModel) GetFlightRecords() (flightRecords []FlightRecord, err error) {
 
 	for rows.Next() {
 		var fr FlightRecord
-		if err = rows.Scan(&fr.UUID, &fr.Date, &fr.MDate, &fr.Departure.Place, &fr.Departure.Time,
+		err = rows.Scan(&fr.UUID, &fr.Date, &fr.MDate, &fr.Departure.Place, &fr.Departure.Time,
 			&fr.Arrival.Place, &fr.Arrival.Time, &fr.Aircraft.Model, &fr.Aircraft.Reg,
 			&fr.Time.SE, &fr.Time.ME, &fr.Time.MCC, &fr.Time.Total, &fr.Landings.Day, &fr.Landings.Night,
 			&fr.Time.Night, &fr.Time.IFR, &fr.Time.PIC, &fr.Time.CoPilot, &fr.Time.Dual, &fr.Time.Instructor,
-			&fr.SIM.Type, &fr.SIM.Time, &fr.PIC, &fr.Remarks); err != nil {
+			&fr.SIM.Type, &fr.SIM.Time, &fr.PIC, &fr.Remarks, &fr.Distance,
+		)
+		if err != nil {
+			return flightRecords, err
+		}
+		m.processFlightrecord(&fr)
+		flightRecords = append(flightRecords, fr)
+	}
+
+	return flightRecords, nil
+}
+
+func (m *DBModel) UpdateFlightRecordTrack(uuid string, distance float64, track []byte) error {
+	ctx, cancel := m.ContextWithDefaultTimeout()
+	defer cancel()
+
+	query := "UPDATE logbook SET distance = ?, track = ? WHERE uuid = ?"
+	_, err := m.DB.ExecContext(ctx, query, distance, track, uuid)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (m *DBModel) GetFlightRecordsForMap() (flightRecords []FlightRecord, err error) {
+	ctx, cancel := m.ContextWithDefaultTimeout()
+	defer cancel()
+
+	rows, err := m.DB.QueryContext(ctx, `
+		SELECT
+			uuid, date, m_date, departure_place, departure_time,
+			arrival_place, arrival_time, aircraft_model, reg_name,
+			se_time, me_time, mcc_time, total_time, day_landings, night_landings,
+			night_time, ifr_time, pic_time, co_pilot_time, dual_time, instructor_time,
+			sim_type, sim_time, pic_name, remarks, distance, track
+		FROM logbook_view
+		ORDER BY m_date desc, departure_time desc`)
+
+	if err != nil {
+		return flightRecords, err
+	}
+	defer rows.Close()
+
+	for rows.Next() {
+		var fr FlightRecord
+		err = rows.Scan(&fr.UUID, &fr.Date, &fr.MDate, &fr.Departure.Place, &fr.Departure.Time,
+			&fr.Arrival.Place, &fr.Arrival.Time, &fr.Aircraft.Model, &fr.Aircraft.Reg,
+			&fr.Time.SE, &fr.Time.ME, &fr.Time.MCC, &fr.Time.Total, &fr.Landings.Day, &fr.Landings.Night,
+			&fr.Time.Night, &fr.Time.IFR, &fr.Time.PIC, &fr.Time.CoPilot, &fr.Time.Dual, &fr.Time.Instructor,
+			&fr.SIM.Type, &fr.SIM.Time, &fr.PIC, &fr.Remarks, &fr.Distance, &fr.Track,
+		)
+		if err != nil {
 			return flightRecords, err
 		}
 		m.processFlightrecord(&fr)