+++ internal/models/helpers.go
@@ -3,32 +3,12 @@ package models
 import (
 	"context"
 	"fmt"
-	"math"
 	"strings"
 	"sync"
 	"time"
+
+	"github.com/vsimakhin/web-logbook/internal/utils"
 )
-
-func deg2rad(degrees float64) float64 {
-	return degrees * math.Pi / 180
-}
-
-func hsin(theta float64) float64 {
-	return math.Pow(math.Sin(theta/2), 2)
-}
-
-// dist calculates a distance between 2 geo points
-func dist(lat1, lon1, lat2, lon2 float64) float64 {
-	lat1 = deg2rad(lat1)
-	lon1 = deg2rad(lon1)
-	lat2 = deg2rad(lat2)
-	lon2 = deg2rad(lon2)
-
-	r := 6378100.0
-	h := hsin(lat2-lat1) + math.Cos(lat1)*math.Cos(lat2)*hsin(lon2-lon1)
-
-	return 2 * r * math.Asin(math.Sqrt(h)) / 1000 / 1.852 // nautical miles
-}
 
 // DefaultTimeout is the default timeout for database queries
 const DefaultTimeout = 60 * time.Second
@@ -46,7 +26,7 @@ func (m *DBModel) ContextWithDefaultTimeout() (context.Context, context.CancelFu
 var distanceCache sync.Map
 
 // distance calculates distance between 2 airports
-func (m *DBModel) distance(departure, arrival string) int {
+func (m *DBModel) Distance(departure, arrival string) float64 {
 	if departure == arrival {
 		return 0
 	}
@@ -59,7 +39,7 @@ func (m *DBModel) distance(departure, arrival string) int {
 	}
 
 	if cachedDistance, ok := distanceCache.Load(key); ok {
-		return cachedDistance.(int)
+		return cachedDistance.(float64)
 	}
 
 	dep, err1 := m.GetAirportByID(departure)
@@ -68,20 +48,57 @@ func (m *DBModel) distance(departure, arrival string) int {
 		return 0
 	}
 
-	d := int(dist(dep.Lat, dep.Lon, arr.Lat, arr.Lon))
+	d := utils.Distance(dep.Lat, dep.Lon, arr.Lat, arr.Lon)
 	distanceCache.Store(key, d)
 	return d
 }
 
 func (m *DBModel) processFlightrecord(fr *FlightRecord) {
-	// calculate distance
-	fr.Distance = m.distance(fr.Departure.Place, fr.Arrival.Place)
-
 	// check for cross country flights
 	if fr.Departure.Place != fr.Arrival.Place {
 		fr.Time.CrossCountry = fr.Time.Total
 	} else {
 		fr.Time.CrossCountry = "0:00"
+	}
+}
+
+func (m *DBModel) UpdateFlightRecordsDistance() {
+	ctx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
+	defer cancel()
+
+	query := "SELECT uuid, departure_place, arrival_place FROM logbook WHERE distance IS NULL"
+	rows, err := m.DB.QueryContext(ctx, query)
+
+	if err != nil {
+		return
+	}
+	defer rows.Close()
+
+	// let's make it in transaction
+	tx, err := m.DB.Begin()
+	if err != nil {
+		return
+	}
+
+	for rows.Next() {
+		var fr FlightRecord
+		if err := rows.Scan(&fr.UUID, &fr.Departure.Place, &fr.Arrival.Place); err != nil {
+			return
+		}
+
+		fr.Distance = m.Distance(fr.Departure.Place, fr.Arrival.Place)
+
+		query = "UPDATE logbook SET distance = ? WHERE uuid = ?"
+		_, err := tx.ExecContext(ctx, query, fr.Distance, fr.UUID)
+		if err != nil {
+			tx.Rollback()
+			return
+		}
+	}
+
+	// commit transaction
+	if err = tx.Commit(); err != nil {
+		return
 	}
 }
 
@@ -107,6 +124,6 @@ func (m *DBModel) CreateDistanceCache() {
 			return
 		}
 
-		m.distance(dep, arr)
+		m.Distance(dep, arr)
 	}
 }