+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+
+- New: Track log support added. You can now upload KML files as attachments or convert existing ones to display flight tracks on the map page. The flight distance is automatically calculated from the track log. Tracks are also visible on the overall Map page.
+
 ## [3.0.0-alpha1] - 17.03.2025
 
 - Caution: This is an early release and still in test, please create a backup of your db file (`weblogbook.sql`) before switching to this version